requirements.txt
MySQL-python==1.2.5
Pillow==2.7.0
boto==2.36.0
image==1.3.4
wsgiref==0.1.2
